Pneumonia-Detection-CNN-main/server/ml_service/requirements.txt
torch
torchvision
timm

opencv-python
numpy
pillow
fastapi
uvicorn
python-multipart
grad-cam
